sasm: modularize code
This is just a quick attempt, `assemble` especially could use more work
but this is mostly just for testing the bootstrap and much of this will
be rewritten anyway.

This is also fixes some bugs I noticed, the `open-output-name` one being
the most egregious as it caused segfaults if no output file was
provided.